"feat: add Class Options to addStudentModal
