+++ src/functions.js
@@ -87,3 +87,17 @@ export class Student {
     this.grades[assignment] = grade;
   }
 }
+
+export function createOptionsFromClasses() {
+  const selectElement = document.createElement("div");
+  if (classes.length != 0) {
+    classes.forEach((classItem) => {
+      const optionElement = document.createElement("option");
+      optionElement.value = classItem.className;
+      optionElement.textContent = classItem.className;
+      selectElement.appendChild(optionElement);
+    });
+  }
+
+  return selectElement;
+}
